Move business logic to store
