+++ src/components/organisms/CesiumViewer.vue
@@ -10,9 +10,9 @@ import Cesium from '@/plugins/cesium';
 import {
   useMap, drawnShape, footprintIds, specifiedShape, visibleOverlayIds,
 } from '@/store';
+import { cesiumViewer, updateTileLayers } from '@/store/cesium';
 import {
-  rgdFootprint, rgdImagery, rgdImageTilesMeta, rgdSpatialEntry, rgdCreateUrl,
-  rgdTokenSignature,
+  rgdFootprint,
 } from '@/api/rest';
 import { RGDResult } from '@/store/types';
 import { Entity } from 'cesium';
@@ -23,18 +23,7 @@ export default defineComponent({
   setup() {
     const polyPoints: number[][][] = [[]];
 
-    const cesiumViewer = ref();
-    let tileSignature: string;
-
-    // const host = rgdHost();
-    // Limit the tile requests on RGD server so that Vue app's requests aren't hung
-    // Cesium.RequestScheduler.requestsByServer = {
-    //   host: 3,
-    // };
-    Cesium.RequestScheduler.maximumRequestsPerServer = 3;
-
     onMounted(async () => {
-      tileSignature = await rgdTokenSignature();
       // Create ProviderViewModel based on different imagery sources
       // - these can be used without Cesium Ion
       const imageryViewModels = [];
@@ -378,54 +367,7 @@ export default defineComponent({
 
     watch(footprintIds, updateFootprints, { deep: true });
 
-    // TODO: this logic needs to be moved to a better place
-    const generateTileProvider = async (imageId: number, band = 0) => {
-      const data = await rgdImageTilesMeta(imageId);
-      const extents = data.bounds;
-      const rectangle = Cesium.Rectangle.fromDegrees(
-        extents.xmin, extents.ymin, extents.xmax, extents.ymax,
-      );
-      const tileProvider = new Cesium.UrlTemplateImageryProvider({
-        url: rgdCreateUrl(`image_process/imagery/${imageId}/tiles/{z}/{x}/{y}.png?projection=EPSG:3857&band=${band}&signature=${tileSignature}`),
-        subdomains: null, // We do not need or provide this in RGD
-        rectangle,
-      });
-      return tileProvider;
-    };
-
-    const tileLayers: Record<string, number> = {}; // Cesium.TileLayer
-    watch(visibleOverlayIds, () => {
-      // Purge
-      Object.keys(tileLayers).forEach(async (key: string) => {
-        const spatialId = Number(key);
-        if (visibleOverlayIds.value.indexOf(spatialId) < 0) {
-          const entry = await rgdSpatialEntry(spatialId);
-          if (entry.subentry_type === 'RasterMeta') {
-            const layers = cesiumViewer.value.scene.imageryLayers;
-            layers.remove(tileLayers[spatialId]);
-            delete tileLayers[spatialId];
-          }
-        }
-      });
-
-      // eslint-disable-next-line no-unused-expressions
-      visibleOverlayIds.value?.forEach(async (spatialId: number) => {
-        if (spatialId in tileLayers) {
-          return;
-        }
-        const entry = await rgdSpatialEntry(spatialId);
-        if (entry.subentry_type === 'RasterMeta') {
-          // TODO: check if present for image ID and Band
-          const imagery = await rgdImagery(spatialId);
-          const imageId = imagery.parent_raster.image_set.images[0].id;
-          const tileProvider = await generateTileProvider(imageId);
-          const layers = cesiumViewer.value.scene.imageryLayers;
-          layers.remove(tileLayers[spatialId]);
-          const tileLayer = layers.addImageryProvider(tileProvider);
-          tileLayers[spatialId] = tileLayer;
-        }
-      });
-    }, { deep: true });
+    watch(visibleOverlayIds, updateTileLayers, { deep: true });
 
     return {
       useMap,
